殡仪馆标准版

+++ src/views/setting/cold.vue
@@ -243,13 +243,13 @@ export default {
       })
     },
     handleDelete(row) {
-      deleteCold(row)
-        .then(res => {
-          this.$confirm('您确认删除吗?', '提示', {
-            confirmButtonText: '确定',
-            cancelButtonText: '取消',
-            type: 'warning'
-          }).then(() => {
+      this.$confirm('您确认删除吗?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteCold(row)
+          .then(res => {
             const index = this.list.indexOf(row)
             this.list.splice(index, 1)
             this.$message({
@@ -262,7 +262,7 @@ export default {
               message: '已取消删除'
             })
           })
-        })
+      })
         .catch(res => {
           this.$notify.error({
             title: '失败',
